Create a shared quantization util for all unit tests.
